Added "baremetal" target for internal RC clocking
Added option for hardware handshaking for UART console:
  -DCONSOLE_RTSCTS

Fixed /bin/passwd and /sbin/mkpasswd to handle correct shadow format and
password offset insertion

Made "TARGET" make variable an override one (?= instead of =) so a target
can be specified at compile time with
  make TARGET=sys/pic32/explorer16

Added "installfs" make target.  This relies on the enabling of an "SDCARD"
variable (or one specified on the command line) containing the path of the
SD card device in order to operate.  Without this variable set the target
fails.  This protects it against unintentional formatting of hard drives