Progress toward a working smallc compiler. Note that this is far from complete, has bugs, and portions of the code generator have not even been looked at yet. However, it can produce working programs from the c code in share/smallc.
To test, cd to share/smallc and type make. Then run ./primelist and/or ./primesum

Smallc is probably a deadend project. It is a very simple compiler that generates code based on a stack oriented model. Mips is pretty much the polar opposite. It can be made to work, but the code size is huge and the execution time will be abysmal. Also, the function calling conventions do not match the standard used for retrobsd so the existing libraries are not simply reusable.

No attempt has been made to generate optimal asm code - the current motivation is simply to generate working asm code.